requirements.txt
python-telegram-bot==20.8
openai==1.54.3
python-dotenv==1.0.0
textblob==0.17.1
pathlib==1.0.1